Avances

+++ app/Http/Controllers/AdmiController.php
@@ -24,8 +24,29 @@ class AdmiController extends Controller
 
 
     public function destroy(CuentaModel $cuenta){
-        $cuenta ->delete();
-        return redirect('Administrador.perfiles');
 
+        if ($cuenta->perfil_id == 1) {
+            return redirect()->route('cuentas.mostrar')->with('error', 'No se puede eliminar al administrador');
+        }
+        $cuenta->delete();
+        return redirect('cuentas.mostrar');
     }
+
+    public function edit(CuentaModel $cuenta){
+        return redirect()->route('cuentas.edit', ['cuenta' => $cuenta]);
+    }
+    
+
+    public function update(Request $request, CuentaModel $cuenta)
+    {
+        $cuenta->nombre = $request->input('nombre');
+        $cuenta->apellido = $request->input('apellido');
+        $cuenta->user = $request->input('user');
+
+
+        $cuenta->save();
+
+        return redirect()->route('cuentas.mostrar')->with('success', 'Usuario actualizado exitosamente');
+    }
+    
 }
